Extract `insert-select-via-input-source` logic from executeQuery.cpp
